+++ R/item_omega.R
@@ -22,10 +22,6 @@
 #' correlation matrix. Required to compute correct fit indices.
 #' @param poly_cor Logical, if `TRUE`, polychoric correlations will be computed
 #' (by passing `poly = TRUE` to `psych::omega()`). Defaults to `FALSE`.
-#' @param item_direction A vector of item keys indicating the direction of
-#' scoring for items. E.g., `c(1, 1, -1, 1)` for four items where the third item
-#' is reverse-scored. Passed to the `key` argument in `psych::omega()`. Defaults
-#' to `NULL`.
 #' @param verbose Logical, if `TRUE` (default), messages are printed.
 #' @param ... Additional arguments passed to [`psych::omega()`].
 #'
@@ -76,7 +72,6 @@ item_omega.data.frame <- function(x,
                                   rotation = "oblimin",
                                   factor_method = "minres",
                                   poly_cor = FALSE,
-                                  item_direction = NULL,
                                   verbose = TRUE,
                                   ...) {
   insight::check_if_installed(c("psych", "parameters"))
@@ -104,7 +99,6 @@ item_omega.data.frame <- function(x,
     rotation = rotation,
     fm = factor_method,
     poly = poly_cor,
-    key = item_direction,
     plot = FALSE,
     ...
   )
@@ -119,7 +113,6 @@ item_omega.data.frame <- function(x,
   attr(out, "rotation") <- rotation
   attr(out, "factor_method") <- factor_method
   attr(out, "poly_cor") <- poly_cor
-  attr(out, "item_direction") <- item_direction
   attr(out, "n") <- n
 
   class(out) <- c("item_omega", "data.frame")
@@ -135,7 +128,6 @@ item_omega.matrix <- function(x,
                               factor_method = "minres",
                               n_obs = NULL,
                               poly_cor = FALSE,
-                              item_direction = NULL,
                               verbose = TRUE,
                               ...) {
   # validate n_obs
@@ -166,7 +158,6 @@ item_omega.matrix <- function(x,
     factor_method = factor_method,
     n.obs = n_obs,
     poly_cor = poly_cor,
-    item_direction = item_direction,
     verbose = verbose,
     ...
   )